+++ tests/test_eventsite.py
@@ -98,6 +98,7 @@ def test_event_site():
                         "id": "submission_deadline",
                         "type": "date",
                     },
+                    "body": {"id": "body", "type": "string"},
                 },
                 "id": "events",
                 "records_api_url": "/type/events/records_api.json",